Rename todo card wrapper, user menu, Fix imports
